requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.2
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-ratelimit==1.0.0
djangorestframework-simplejwt==5.3.1
Markdown==3.6
pillow==10.4.0
PyJWT==2.8.0
pymemcache==4.0.0
python-memcached==1.62
sqlparse==0.5.1
tzdata==2024.1
